Cast status as strings.
This is to aid one of the most idiotic lines of code
I've seen and which is included in Textpattern 4.5.2.

This bug in 4.5.2 makes the status appear as zeros
in the list. This is due to 4.5.2 casting all
options to the default/selected options type.

Since the default value is an integer, the presented
labels are converted to '0' (zeros).